+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Removed
+- Remove deprecated `not_equal` syntax in favor of `not-equal`.
+
 
 ## 0.3.5 (2018-04-11)
 ### Added
@@ -13,7 +16,6 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Help documentation for flags with placeholder options now display them using
   angular brackets.
 - Minor changes to certain error messages.
-
 
 ## 0.3.4 (2018-03-14)
 ### Added